RacingTest에서 Car에 대한 테스트케이스 분리
