+++ src/test/java/mission1/racing/RacingTest.java
@@ -12,7 +12,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 public class RacingTest {
     RacingRandom random = new RacingRandom();
     RacingGame racingGame = new RacingGame();
-    Car car = new Car(5);
 
     @DisplayName("랜덤의 범위가 0~9인지 확인")
     @Test
@@ -58,21 +57,6 @@ public class RacingTest {
         }
 
         assertEquals(racingGame.getWinnersArrayList().size(), numberOfWinners);
-    }
-
-    @DisplayName("우승자를 제대로 가져오는지 확인 - 파라미터와 동일한 경우 true 반환")
-    @ParameterizedTest
-    @CsvSource(value = {"1 : false", "3 : false", "5 : true", "6 : false", "0 : false"}, delimiter = ':')
-    void testGetWinner(int maxPosition, boolean expected) {
-        assertEquals(car.isWinner(maxPosition), expected);
-    }
-
-    @DisplayName("자동차의 이동이 제대로 이뤄지는지 확인 - 파라미터가 4 이상일때 carPosition 증가 ")
-    @ParameterizedTest
-    @CsvSource(value = {"0, 5", "1, 5", "3, 5", "4, 6", "5, 6", "6, 6"}, delimiter = ',')
-    void testCarGo(int randomValue, int expectedPosition) {
-        car.go(randomValue);
-        assertEquals(car.getCarPositions(), expectedPosition);
     }
 }
